Restructuring file tree.
