+++ src/components/Navigation/SideDrawer/SideDrawer.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Logo from '../../Logo/Logo'
 import NavigationItems from '../NavigationItems/NavigationItems'
 import Backdrop from '../../UI/Backdrop/Backdrop'
-import Aux from '../../../hoc/Auxiliary'
+import Aux from '../../../hoc/Auxiliary/Auxiliary'
 import styles from './SideDrawer.module.css'
 
 //? 'show' is a prop of <Backdrop /> that store a bool so no
