+++ src/components/OrderSummary/OrderSummary.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Aux from '../../hoc/Auxiliary'
+import Aux from '../../hoc/Auxiliary/Auxiliary'
 import Button from '../UI/Button/Button'
 
 //! With functional components, you only need to use {} if
