+++ src/hoc/Layout/Layout.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Aux from '../Auxiliary'
+import Aux from '../Auxiliary/Auxiliary'
 import styles from './Layout.module.css'
 import Toolbar from '../../components/Navigation/Toolbar/Toolbar'
 import SideDrawer from '../../components/Navigation/SideDrawer/SideDrawer'
